backend/requirements.txt
# FastAPI Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0

# AI/ML
torch>=2.2.0
transformers>=4.36.2
Pillow>=10.2.0
numpy>=1.26.0
sentence-transformers>=2.3.1

# Telegram
python-telegram-bot==20.8

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
aiofiles==23.2.1

# Image Processing
opencv-python-headless==4.9.0.80

# API Client
httpx==0.26.0
requests==2.31.0
